+++ packages/core/create-cable/index.d.ts
@@ -4,10 +4,10 @@ import { Encoder } from '../encoder/index.js'
 import { Monitor, ReconnectStrategy } from '../monitor/index.js'
 import { Cable } from '../cable/index.js'
 import { Protocol } from '../protocol/index.js'
-import { Channel, Message, ChannelParamsMap } from '../channel/index.js'
 import { Options } from '../action_cable/index.js'
 import { ExtendedOptions } from '../action_cable_ext/index.js'
 import { WebSocketAuthStrategy } from '../websocket/index.js'
+import { ActionCableConsumer } from '../action_cable_compat/index.js'
 
 export type ExtendedProtocolID =
   | 'actioncable-v1-ext-json'
@@ -82,43 +82,8 @@ export function createCable(
   opts?: Partial<CreateOptions<ProtocolID | Protocol>>
 ): Cable
 
-export type ActionCableMixin<T extends Message> = Partial<{
-  initialized: () => void
-  connected: () => void
-  rejected: () => void
-  received: (data: T) => void
-  disconnected: () => void
-}>
-
-export class ActionCableSubscription {
-  channel: ActionCableChannel
-  identifier: string
-
-  perform(action: string, payload?: object): void
-  send(payload: object): void
-  unsubscribe(): void
-}
-
-export class ActionCableChannel extends Channel {
-  subscription: ActionCableSubscription
-}
-
-export class ActionCableSubscriptions {
-  create<M extends ActionCableMixin<T> = {}, T extends Message = Message>(
-    params: ChannelParamsMap | string,
-    mixin: ActionCableMixin<T>
-  ): ActionCableSubscription & M
-  create(params: ChannelParamsMap | string): ActionCableSubscription
-  findAll(identifier: string): ActionCableSubscription[]
-}
-
-export class ActionCableConsumer {
-  readonly cable: Cable
-  readonly subscriptions: ActionCableSubscriptions
-}
-
 export function createConsumer(
-  url: string,
+  url?: string,
   opts?: Partial<CreateOptions<ProtocolID | Protocol>>
 ): ActionCableConsumer
 export function createConsumer(